Drop ExoPlayer specific exceptions in favour of NoPlayerError.
