+++ core/src/main/java/com/novoda/noplayer/internal/exoplayer/forwarder/ExoPlayerErrorMapper.java
@@ -4,14 +4,10 @@ import android.media.MediaCodec;
 
 import com.google.android.exoplayer2.ParserException;
 import com.google.android.exoplayer2.upstream.HttpDataSource;
+import com.novoda.noplayer.NoPlayerError;
 import com.novoda.noplayer.Player;
+import com.novoda.noplayer.PlayerErrorType;
 import com.novoda.noplayer.drm.StreamingModularDrm;
-import com.novoda.noplayer.internal.exoplayer.playererror.ConnectivityError;
-import com.novoda.noplayer.internal.exoplayer.playererror.DrmDecryptionError;
-import com.novoda.noplayer.internal.exoplayer.playererror.DrmRequestError;
-import com.novoda.noplayer.internal.exoplayer.playererror.InvalidResponseCodeError;
-import com.novoda.noplayer.internal.exoplayer.playererror.MalformedContentError;
-import com.novoda.noplayer.internal.exoplayer.playererror.UnknownError;
 
 import java.io.IOException;
 
@@ -23,22 +19,25 @@ final class ExoPlayerErrorMapper {
 
     static Player.PlayerError errorFor(Exception e) {
         if (e instanceof HttpDataSource.InvalidResponseCodeException) {
-            return new InvalidResponseCodeError(e);
+            return new NoPlayerError(PlayerErrorType.INVALID_RESPONSE_CODE, e);
         }
+
         if (e instanceof ParserException) {
-            return new MalformedContentError(e);
+            return new NoPlayerError(PlayerErrorType.MALFORMED_CONTENT, e);
         }
 
         Throwable cause = e.getCause();
         if (e.getCause() instanceof MediaCodec.CryptoException) {
-            return new DrmDecryptionError(e);
+            return new NoPlayerError(PlayerErrorType.FAILED_DRM_DECRYPTION, e);
         }
+
         if (cause instanceof StreamingModularDrm.DrmRequestException) {
-            return new DrmRequestError(cause);
+            return new NoPlayerError(PlayerErrorType.FAILED_DRM_REQUEST, e);
         }
+
         if (e instanceof IOException || cause instanceof IOException) {
-            return new ConnectivityError(cause == null ? e : cause);
+            return new NoPlayerError(PlayerErrorType.CONNECTIVITY_ERROR, cause == null ? e : cause);
         }
-        return new UnknownError(e);
+        return new NoPlayerError(PlayerErrorType.UNKNOWN, e);
     }
 }
