+++ core/src/main/java/com/novoda/noplayer/internal/exoplayer/forwarder/MediaSourceOnErrorForwarder.java
@@ -1,8 +1,9 @@
 package com.novoda.noplayer.internal.exoplayer.forwarder;
 
 import com.google.android.exoplayer2.source.ExtractorMediaSource;
+import com.novoda.noplayer.NoPlayerError;
 import com.novoda.noplayer.Player;
-import com.novoda.noplayer.internal.exoplayer.playererror.MediaSourceError;
+import com.novoda.noplayer.PlayerErrorType;
 
 import java.io.IOException;
 
@@ -16,7 +17,7 @@ class MediaSourceOnErrorForwarder implements ExtractorMediaSource.EventListener 
 
     @Override
     public void onLoadError(IOException error) {
-        Player.PlayerError playerError = new MediaSourceError(error);
+        Player.PlayerError playerError = new NoPlayerError(PlayerErrorType.MEDIA_SOURCE_ERROR, error);
         errorListener.onError(playerError);
     }
 }
