--- core/src/main/java/com/novoda/noplayer/internal/exoplayer/playererror/ConnectivityError.java
@@ -1,22 +1,0 @@
-package com.novoda.noplayer.internal.exoplayer.playererror;
-
-import com.novoda.noplayer.Player;
-
-public class ConnectivityError implements Player.PlayerError {
-
-    private final Throwable cause;
-
-    public ConnectivityError(Throwable cause) {
-        this.cause = cause;
-    }
-
-    @Override
-    public String getType() {
-        return "ConnectivityError";
-    }
-
-    @Override
-    public Throwable getCause() {
-        return cause;
-    }
-}
