--- core/src/main/java/com/novoda/noplayer/internal/exoplayer/playererror/DrmDecryptionError.java
@@ -1,22 +1,0 @@
-package com.novoda.noplayer.internal.exoplayer.playererror;
-
-import com.novoda.noplayer.Player;
-
-public class DrmDecryptionError implements Player.PlayerError {
-
-    private final Throwable cause;
-
-    public DrmDecryptionError(Throwable cause) {
-        this.cause = cause;
-    }
-
-    @Override
-    public String getType() {
-        return "DrmDecryptionError";
-    }
-
-    @Override
-    public Throwable getCause() {
-        return cause;
-    }
-}
