--- core/src/main/java/com/novoda/noplayer/internal/exoplayer/playererror/InvalidResponseCodeError.java
@@ -1,22 +1,0 @@
-package com.novoda.noplayer.internal.exoplayer.playererror;
-
-import com.novoda.noplayer.Player;
-
-public class InvalidResponseCodeError implements Player.PlayerError {
-
-    private final Throwable cause;
-
-    public InvalidResponseCodeError(Throwable cause) {
-        this.cause = cause;
-    }
-
-    @Override
-    public String getType() {
-        return "InvalidResponseCodeError";
-    }
-
-    @Override
-    public Throwable getCause() {
-        return cause;
-    }
-}
