--- core/src/main/java/com/novoda/noplayer/internal/exoplayer/playererror/MalformedContentError.java
@@ -1,22 +1,0 @@
-package com.novoda.noplayer.internal.exoplayer.playererror;
-
-import com.novoda.noplayer.Player;
-
-public class MalformedContentError implements Player.PlayerError {
-
-    private final Throwable cause;
-
-    public MalformedContentError(Throwable cause) {
-        this.cause = cause;
-    }
-
-    @Override
-    public String getType() {
-        return "MalformedContentError";
-    }
-
-    @Override
-    public Throwable getCause() {
-        return cause;
-    }
-}
