--- core/src/main/java/com/novoda/noplayer/internal/exoplayer/playererror/MediaSourceError.java
@@ -1,22 +1,0 @@
-package com.novoda.noplayer.internal.exoplayer.playererror;
-
-import com.novoda.noplayer.Player;
-
-public class MediaSourceError implements Player.PlayerError {
-
-    private final Throwable cause;
-
-    public MediaSourceError(Throwable cause) {
-        this.cause = cause;
-    }
-
-    @Override
-    public String getType() {
-        return "MediaSourceError";
-    }
-
-    @Override
-    public Throwable getCause() {
-        return cause;
-    }
-}
