--- core/src/main/java/com/novoda/noplayer/internal/exoplayer/playererror/UnknownError.java
@@ -1,22 +1,0 @@
-package com.novoda.noplayer.internal.exoplayer.playererror;
-
-import com.novoda.noplayer.Player;
-
-public class UnknownError implements Player.PlayerError {
-
-    private final Throwable cause;
-
-    public UnknownError(Throwable cause) {
-        this.cause = cause;
-    }
-
-    @Override
-    public String getType() {
-        return "UnknownError";
-    }
-
-    @Override
-    public Throwable getCause() {
-        return cause;
-    }
-}
